feat: Expose `pint build` command args for use downstream
This enables the downstream `pint-deploy` tool to also receive uniform
"build" arguments that match what the `pint build` command expects.
